format test_parser and add conftest
