+++ tests/conftest.py
@@ -1,0 +1,16 @@
+import pytest
+
+from solaudit.models import Program
+from solaudit.parsers import getProgramParser, comment
+
+
+@pytest.fixture
+def program():
+    return Program()
+
+
+@pytest.fixture
+def parser(program):
+    parser = getProgramParser(program)
+    parser.ignore(comment)
+    return parser
