+++ tests/test_checkers.py
@@ -7,18 +7,6 @@ from solaudit.models import Program
 from solaudit.parsers import getProgramParser, comment
 
 
-@pytest.fixture
-def program():
-    return Program()
-
-
-@pytest.fixture
-def parser(program):
-    parser = getProgramParser(program)
-    parser.ignore(comment)
-    return parser
-
-
 class TestCheckers:
     def test_overUnderFlowChecker(self, program, parser):
         content = """
